% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_barplot.R
\name{create_head}
\alias{create_head}
\title{Produce a barplot (geom_col) from the summary_table}
\usage{
create_head(subdata)
}
\arguments{
\item{subset}{data frame with variables necessary for scatterplot}
}
\value{
tibble object for the scatterplot
}
\description{
Produce a barplot (geom_col) from the summary_table
}
\examples{
plot_obj <- create_barplot(subdata)
plot_obj

}
