% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summary.R
\name{create_summary}
\alias{create_summary}
\title{Produce a summary table for the 3 variables}
\usage{
create_summary(subdata)
}
\arguments{
\item{subset}{data frame with variables necessary for scatterplot}
}
\value{
tibble object for the scatterplot
}
\description{
Produce a summary table for the 3 variables
}
\examples{
summary_table <- scatter_sales(data = ames, xvar = "Lot_Frontage", yvar = "Sale_Price")
plot_obj
}
